+++ .gitignore
@@ -14,6 +14,7 @@ release.*
 __pycache__/
 *.py[cod]
 *$py.class
+DirectoryStructure.java
 
 # C extensions
 *.so
