+++ modular_rl/agents/_common/__init__.py
@@ -1,2 +1,1 @@
 from ._agents import CommonAgents
-from .mcis import CommonMCIS
