+++ modular_rl/agents/mcis.py
@@ -3,6 +3,9 @@ from modular_rl.agents.tensorflow.mcis import TensorFlowAgentMCIS
 
 
 class AgentMCIS():
+    LEARN_PYTORCH = "pytorch"
+    LEARN_TENSORFLOW = "tensorflow"
+
     def __init__(self, env, setting):
         """
         Initialize the AgentMCIS class with the specified environment and settings.
@@ -12,8 +15,52 @@ class AgentMCIS():
         :param setting: The settings for the MCIS algorithm.
         :type setting: AgentSettings
         """
+        self.env = env
+        self.setting = setting
 
-        if setting['learn_framework'] == 'pytorch':
+        learn_framework = self.setting.get(
+            "learn_framework", self.LEARN_PYTORCH)
+
+        if learn_framework == self.LEARN_PYTORCH:
             self.agent = PyTorchAgentMCIS(env, setting)
-        elif setting['learn_framework'] == 'tensorflow':
+        elif learn_framework == self.LEARN_TENSORFLOW:
             pass
+
+    def select_action(self, state):
+        return self.agent.select_action(state)
+
+    def backpropagate(self, search_path, reward, done):
+        self.agent.backpropagate(search_path, reward, done)
+
+    def learn(self):
+        self.agent.learn()
+
+    def train(self):
+        self.agent.train()
+
+    def compute_loss(self, state, action, reward, next_state, done):
+        return self.agent.compute_loss(state, action, reward, next_state, done)
+
+    def update(self):
+        self.agent.update()
+
+    def check_tensor(self, state):
+        return self.agent.check_tensor(state)
+
+    def save_model(self, file_name):
+        self.agent.save_model(file_name)
+
+    def save(self, file_name):
+        self.agent.save(file_name)
+
+    def load_model(self, file_name):
+        self.agent.load_model(file_name)
+
+    def load(self, file_name):
+        self.agent.load(file_name)
+
+    def reset(self):
+        self.agent.reset()
+
+    def learn_reset(self):
+        self.agent.learn_reset()
