+++ modular_rl/agents/pytorch/__init__.py
@@ -1,4 +1,3 @@
 from .mcis import PyTorchAgentMCIS
-from .mcts import AgentMCTS
-from .ppo import AgentPPO
-from .mim import AgentMIM
+from .mcts import PyTorchAgentMCTS
+from .ppo import PyTorchAgentPPO
