+++ modular_rl/agents/pytorch/mcts.py
@@ -1,61 +1,69 @@
 
+# -*- coding: utf-8 -*-
 """
 ModularRL project
 
 Copyright (c) 2023 horrible-gh
 
-Class AgentMCIS is an implementation of the Monte Carlo Importance Sampling (MCIS) algorithm.
-The algorithm is used for solving problems of sequential decision making under uncertainty.
+Class AgentMCTS is an implementation of the Monte Carlo Tree Search (MCTS) algorithm.
 It takes an environment and a setting configuration as inputs, initializes neural network instances and optimizers,
 and sets various learning parameters.
-This class has methods to predict an action given a state, perform a learning step, update the neural network parameters,
+It has methods to predict an action given a state, perform a learning step, update the neural network parameters,
 save and load a checkpoint, and reset learning parameters.
 The class also has instance variables to keep track of episode and total rewards, previous reward, and average reward.
-
-Importance Sampling is used to estimate the properties of a particular target distribution, given some observed data and a proposal distribution.
-This implementation makes use of Monte Carlo methods, which rely on repeated random sampling to obtain numerical results.
 
 This software includes the following third-party libraries:
 Gym (MIT License): https://github.com/openai/gym - Copyright (c) OpenAI.
 NumPy (BSD License): https://numpy.org - Copyright (c) NumPy Developers.
-PyTorch (BSD-Style License): https://pytorch.org/ - Copyright (c) Facebook.
+PyTorch  (BSD-Style License): https://pytorch.org/ - Copyright (c) Facebook.
 """
 
 import torch
+import numpy as np
 from torch.distributions import Categorical
 import torch.nn.functional as F
 from modular_rl.util.node import Node
-from modular_rl.agents.pytorch._agent import Agent
-import numpy as np
+from modular_rl.agents._agent import Agent
+from LogAssist.log import Logger
 
 
-class PyTorchAgentMCIS(Agent):
+class PyTorchAgentMCTS(Agent):
     def __init__(self, env, setting):
         """
-        Initialize the AgentMCIS class with the specified environment and settings.
+        Initialize the AgentMCTS class with the specified environment and settings.
 
         :param env: The environment to use for training.
         :type env: gym.Env or None
-        :param setting: The settings for the MCIS algorithm.
+        :param setting: The settings for the MCTS algorithm.
         :type setting: AgentSettings
         """
 
-        super(PyTorchAgentMCIS, self).__init__(env, setting)
-        super(PyTorchAgentMCIS, self).init_actor_critic()
+        super(PyTorchAgentMCTS, self).__init__(env, setting)
+        super(PyTorchAgentMCTS, self).init_actor_critic()
 
-        # mcis parameters
+        # MCTS parameters
         self.num_simulations = setting.get('num_simulations', 800)
         self.cpuct = setting.get('cpuct', 1.0)
         self.temperature = setting.get('temperature', 1.0)
+        self.gamma = 0.95
+        self.total_value = 0
 
         self.device = setting.get('device', None)
         if self.device == None:
             self.device = torch.device(
                 "cuda" if torch.cuda.is_available() else "cpu")
+        self.reset()
+
+    def reset(self):
+        self.state = None
+        self.action = None
+        self.reward = None
+        self.done = None
+        self.total_reward = 0
 
     def select_action(self, state):
         """
-        Select an action using mcis.
+        Select an action using MCTS.
 
         :param state: The current state.
         :type state: numpy.ndarray
@@ -63,34 +71,50 @@ class PyTorchAgentMCIS(Agent):
         :rtype: int
         """
 
-        state_tensor = self.check_tensor(state).to(self.device)
+        state_tensor = self.check_tensor(self._check_state(state))
         action_probs, _ = self.actor_critic_net(state_tensor)
-        action_probs = action_probs.detach().numpy().flatten()
-        root = Node(state, action_probs)
+        action_probs = action_probs.detach()
+        m = Categorical(action_probs)
+        chosen_action = m.sample().item()
+
+        # Use uniform prior for root node.
+        root = Node(state, [1 / self.env.action_space.n]
+                    * self.env.action_space.n)
+
         for _ in range(self.num_simulations):
-            node = root
-            search_path = [node]
+            node, search_path = root, [root]
+
             while node.expanded():
                 action, node = node.select_child(self.cpuct)
+                if action is None:
+                    action = chosen_action
                 search_path.append(node)
-            if len(search_path) > 1:
-                parent, action = search_path[-2], search_path[-1].action
-            else:
-                parent, action = search_path[0], None
-            if True not in self.dones:  # Check if the game is not over
-                state_tensor = self.check_tensor(node.state).to(self.device)
-                action_probs, value = self.actor_critic_net(state_tensor)
-                action_space = self.env.action_space.n
-                node.expand(action_space, action_probs, False)
-            state_tensor = self.check_tensor(node.state).to(self.device)
-            _, value = self.actor_critic_net(state_tensor)
-            self.backpropagate(search_path, value.item(), node.done)
-        root_state_tensor = self.check_tensor(
-            root.state).to(self.device)  # This line is added
-        action_probs, _ = self.actor_critic_net(root_state_tensor)
-        chosen_action = np.random.choice(
-            range(len(action_probs)), p=action_probs.detach().numpy())
-        return chosen_action
+
+            parent, action = (search_path[-2], search_path[-1].action) if len(
+                search_path) > 1 else (search_path[0], None)
+            step_output = self.env.step(action) if action is not None else (
+                parent.state, 0, False, None)
+            state, reward, done, *_ = step_output
+            Logger.verb('mcts:select_action', f"Step Output Reward: {reward}")
+
+            if not done:
+                state_tensor = self.check_tensor(state).to(self.device)
+                action_probs, _ = self.actor_critic_net(state_tensor)
+                node.expand(self.env.action_space.n,
+                            action_probs.detach().cpu().numpy().flatten(), False)
+
+            self.backpropagate(search_path, reward, done)
+
+        chosen_action_node = root.children[chosen_action]
+
+        self.state = chosen_action_node.state
+        self.action = chosen_action
+        self.reward = chosen_action_node.total_value
+        self.done = self.reward != 0
+
+        self.total_reward += self.reward
+
+        return self.state, self.action, self.reward, self.done
 
     def backpropagate(self, search_path, reward, done):
         """
@@ -104,18 +128,9 @@ class PyTorchAgentMCIS(Agent):
         :type done: bool
         """
         for node in reversed(search_path):
+            # Logger.verb('mcts:backpropagate',
+            #            f"Node Reward({reward}) after Update Stats: {node.total_value}")
             node.update_stats(reward)
-            if not done:
-                if isinstance(node.state, np.ndarray):
-                    state_tensor = torch.from_numpy(
-                        node.state).float().to(self.device)
-                elif torch.is_tensor(node.state):
-                    state_tensor = node.state.float().to(self.device)
-                else:
-                    raise ValueError(
-                        "node.state must be a numpy array or torch tensor")
-                _, reward = self.actor_critic_net(state_tensor)
-                reward = reward.item()
 
     def learn(self):
         """
@@ -131,32 +146,28 @@ class PyTorchAgentMCIS(Agent):
         for episode in range(self.max_episodes):
             self.episode = episode
             state = self.learn_reset()
+            # state = self.env.reset()
 
             for t in range(self.max_timesteps):
                 state = self._check_state(state)
                 self.state_tensor = self.check_tensor(state).squeeze(0)
+                self.next_state, self.action, self.reward, self.done = self.select_action(
+                    self.state_tensor)
 
-                # Select an action
-                action = self.select_action(self.state_tensor)
+                self.learn_check()
 
-                # Take a step in the environment
-                step_out = self.env.step(action)
-                next_state, reward, done = self.step_unpack(step_out)
-
-                # Update the agent's experience
-                self.update_step(state, action, reward, done, next_state)
-
-                # Update the network parameters at the end of the episode
-                if done:
-                    self.update()
-                    self.learn_check()
+                if self.done:
+                    # self.update()
                     break
 
-                state = next_state
+                self.update()
+
+                state = self.next_state
 
             self.episode += 1
+            # print(f"Episode: {self.episode}, Reward: {self.total_reward}")
 
-    def compute_loss(self, state, action, reward, next_state, done):
+    def compute_loss(self, state, action, reward, done):
         '''
         This function computes the actor and critic loss using the provided state, action, reward, next_state, and done variables.
         The actor loss is computed based on the policy gradient algorithm,
@@ -177,64 +188,61 @@ class PyTorchAgentMCIS(Agent):
         :param done: A flag indicating whether the episode has ended.
         :return: The computed actor and critic loss values.
         '''
-
         # Predict action probabilities and values
+        state = self.ensure_float(state)
         action_probs, values = self.actor_critic_net(state)
 
-        # Logger.verb('agents:mcis:compute_loss',f'{self.actor_critic_net(next_state)}')
-
         # Compute the value loss
-        actor_output, critic_output = self.actor_critic_net(next_state)
-        target_values = reward + self.gamma * \
-            torch.mean(critic_output) * (1 - done)
-        target_values = target_values.unsqueeze(1)
+        # Convert reward to tensor
+        reward = torch.tensor(reward, device=self.device)
+        # Convert total_value to tensor
+        total_value = torch.tensor(self.total_value, device=self.device)
+        target_values = reward + self.gamma * total_value * (1 - done)
+        target_values = target_values.unsqueeze(0)
         critic_loss = F.mse_loss(values, target_values.detach())
+
+        # If action is not a list or tuple or its length is zero, initialize it with zeros
+        if not isinstance(action, (list, tuple)) or len(action) == 0:
+            action = torch.zeros_like(action_probs)
 
         # Compute the policy loss
         m = Categorical(action_probs)
         logprobs = m.log_prob(self.check_tensor(action))
         actor_loss = -logprobs * (target_values - values).detach()
 
-        # Take mean of actor_loss and critic_loss
-        actor_loss = actor_loss.mean()
-        critic_loss = critic_loss.mean()
+        # Average the actor and critic loss
+        loss = (actor_loss.mean() + critic_loss.mean()) / 2
 
-        return actor_loss, critic_loss
+        return loss
 
     def update(self):
-        """
+        '''
         This function updates the network parameters using the optimizer and computed loss values.
-        """
 
-        if not self.states:  # Check if the states list is empty
-            return
+        update() function updates the network parameters using the optimizer and computed loss values.
+        It uses the compute_loss() function to compute the loss and the optimizer object to perform the optimization step.
 
-        # Logger.verb('agents:mcis:update',f'states={self.states},actions={self.actions},rewards={self.rewards},dones={self.dones}')
+        This function does not take any parameters and does not return anything.
 
-        # Prepare data
-        states_tensor = torch.stack(
-            [self.check_tensor(state) for state in self.states]).to(self.device)
-        actions_tensor = torch.Tensor(self.actions).to(self.device)
-        rewards_tensor = torch.Tensor(self.rewards).to(self.device)
-        next_states_tensor = torch.stack(
-            [self.check_tensor(state) for state in self.next_states]).to(self.device)
-        dones_tensor = torch.Tensor(self.dones).to(self.device)
+        :return: None
+        '''
 
-        # Compute loss
-        actor_loss, critic_loss = self.compute_loss(
-            states_tensor, actions_tensor, rewards_tensor, next_states_tensor, dones_tensor)
+        # Update the network
+        # Convert state to tensor
+        state_tensor = torch.tensor(self.state, device=self.device)
 
-        # Compute gradients and update network parameters
+        # Update the network
         self.actor_critic_optimizer.zero_grad()
-        loss = actor_loss + critic_loss
+        loss = self.compute_loss(
+            state_tensor, self.action, self.reward, self.done)
+        Logger.verb('mcts:update',
+                    f"Loss: {loss.item()}, Reward: {self.reward}")
         loss.backward()
         self.actor_critic_optimizer.step()
 
-        # Reset the lists for the next episode
         self.reset()
-        self.update_episode()
 
-    def check_tensor(self, state):
+    def check_tensor(self, obj):
         '''
         This function checks if the provided object is a PyTorch tensor, and if not, converts it to a tensor.
 
@@ -250,9 +258,18 @@ class PyTorchAgentMCIS(Agent):
         :return: The input object as a PyTorch tensor.
         '''
 
-        if not isinstance(state, torch.Tensor):
-            state = torch.tensor(state, dtype=torch.float32)
-        return state
+        if not torch.is_tensor(obj):
+            obj_tensor = torch.FloatTensor(obj)
+        else:
+            obj_tensor = obj
+
+        Logger.verb('mcts:check_tensor', f"Before remapping: {obj_tensor}")
+        obj_tensor = obj_tensor % self.env.action_space.n
+        obj_tensor = torch.where(
+            obj_tensor >= self.env.action_space.n, self.env.action_space.n - 1, obj_tensor)
+        obj_tensor = torch.round(obj_tensor)
+        Logger.verb('mcts:check_tensor', f"After remapping: {obj_tensor}")
+        return obj_tensor
 
     def save_model(self, file_name):
         """
@@ -291,3 +308,17 @@ class PyTorchAgentMCIS(Agent):
         """
 
         self.load_actor_critic(file_name)
+
+    def update_step(self, state, action, reward, done, next_state):
+        """
+        In MCTS, update_step is not necessary as the agent is updated
+        after every simulation (or a batch of simulations), each of which
+        represents an entire episode. Thus, this method can be left blank or removed.
+        """
+        self.state = state
+        self.action = action
+        self.reward = reward
+        self.done = done
+        self.next_state = next_state
+        self.update_reward(reward)
+        self.update_episode()
