+++ modular_rl/agents/tensorflow/_agent.py
@@ -1,4 +1,5 @@
 import tensorflow as tf
+from tensorflow.keras.models import load_model
 from LogAssist.log import Logger
 from modular_rl.networks.tensorflow.policy import PolicyNetwork
 from modular_rl.networks.tensorflow.value import ValueNetwork
@@ -65,3 +66,66 @@ class Agent(CommonAgents):
             self.state_dim, self.action_dim)
         self.actor_critic_optimizer = tf.keras.optimizers.Adam(
             learning_rate=self.setting.get('optimizer_speed', 3e-4))
+
+    def select_action(self, state):
+        '''
+        These functions are placeholders and must be implemented by the child class that extends this Agent class.
+
+        select_action() function is a placeholder that needs to be implemented in the child class that extends the Agent class. This function takes the current state of the environment and returns the selected action for the agent to take.
+
+        :param state: The current state of the environment.
+        :return: The selected action for the agent to take.
+        '''
+        pass
+
+    def update(self):
+        '''
+        This function is a placeholder and must be implemented by the child class that extends this Agent class.
+
+        update() function is a placeholder that needs to be implemented in the child class that extends the Agent class. This function is responsible for updating the agent's state, action, and policy based on the new state and reward received from the environment.
+
+        No parameters are passed into this function and it does not return anything.
+        '''
+        pass
+
+    def save_policy_value(self, file_name):
+        """
+        Save the policy and value networks.
+
+        :param file_name: The base name of the files to save the networks.
+        :type file_name: str
+        """
+
+        self.policy_net.save(file_name + '_policy.h5')
+        self.value_net.save(file_name + '_value.h5')
+
+    def load_policy_value(self, file_name):
+        """
+        Load the policy and value networks.
+
+        :param file_name: The base name of the files to load the networks.
+        :type file_name: str
+        """
+
+        self.policy_net = load_model(file_name + '_policy.h5')
+        self.value_net = load_model(file_name + '_value.h5')
+
+    def save_actor_critic(self, file_name):
+        """
+        Save the actor-critic network.
+
+        :param file_name: The name of the file to save the network.
+        :type file_name: str
+        """
+
+        self.actor_critic_net.save(file_name + '_actor_critic.h5')
+
+    def load_actor_critic(self, file_name):
+        """
+        Load the actor-critic network.
+
+        :param file_name: The name of the file to load the network.
+        :type file_name: str
+        """
+
+        self.actor_critic_net = load_model(file_name + '_actor_critic.h5')
