+++ modular_rl/networks/pytorch/__init__.py
@@ -1,3 +1,3 @@
-from .policy import PolicyNetwork
-from .value import ValueNetwork
-from .actor_critic import ActorCriticNetwork
+from .policy import PyTorchPolicyNetwork
+from .value import PyTorchValueNetwork
+from .actor_critic import PyTorchActorCriticNetwork
