+++ modular_rl/networks/pytorch/policy.py
@@ -4,14 +4,14 @@ ModularRL project
 
 Copyright (c) 2023 horrible
 
-PolicyNetwork is a neural network that takes in the current state of the environment as input and outputs a probability distribution over the available actions. 
+PolicyNetwork is a neural network that takes in the current state of the environment as input and outputs a probability distribution over the available actions.
 The purpose of this network is to determine the best action to take in the current state.
 
-ValueNetwork, on the other hand, 
-is a neural network that takes in the current state of the environment as input and outputs a value that represents the expected future reward that can be obtained from that state. 
+ValueNetwork, on the other hand,
+is a neural network that takes in the current state of the environment as input and outputs a value that represents the expected future reward that can be obtained from that state.
 The purpose of this network is to estimate the value of a given state so that the agent can make decisions that lead to the most reward in the long term.
 
-Both networks are important components of the Proximal Policy Optimization (PPO) algorithm, 
+Both networks are important components of the Proximal Policy Optimization (PPO) algorithm,
 which is a type of reinforcement learning algorithm used for training agents to perform tasks in an environment.
 
 This software includes the following third-party libraries:
@@ -22,7 +22,7 @@ import torch.nn as nn
 import torch
 
 
-class PolicyNetwork(nn.Module):
+class PyTorchPolicyNetwork(nn.Module):
     def __init__(self, input_dim, output_dim, hidden_option='medium', num_layers=2):
         """
         Initialize the PolicyNetwork class with the specified input dimension, output dimension, and network settings.
@@ -37,7 +37,7 @@ class PolicyNetwork(nn.Module):
         :type num_layers: int
         """
 
-        super(PolicyNetwork, self).__init__()
+        super(PyTorchPolicyNetwork, self).__init__()
 
         hidden_dim_map = {
             'small': 32,
