+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "modular_rl"
-version = "0.3.3"
+version = "0.4.0"
 authors = [
   { name="horrible", email="shinjpn1@gmail.com" },
 ]
@@ -23,7 +23,7 @@ classifiers = [
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
 ]
-keywords=['ModularRL', 'Modular-RL', 'Modular_RL', 'Modular RL',  'modularrl',  'modular_rl', 'modular-rl', 'modular rl', 'mrl', 'modular', 'learn', 'learning', 'pytorch learn', 'pytorch learning']
+keywords=['ModularRL', 'Modular-RL', 'Modular_RL', 'Modular RL',  'modularrl',  'modular_rl', 'modular-rl', 'modular rl', 'mrl', 'modular', 'learn', 'learning', 'pytorch learn', 'pytorch learning', 'tensorflow learn', 'tensorflow learning', 'ai']
 dependencies=['gym', 'numpy', 'torch', 'LogAssist']
